Net6 upgrade (#79)
* Update projects to Net6
* Fix build errors

+++ src/PlcInterface.OpcUa/Extensions/SessionExtensions.cs
@@ -77,7 +77,7 @@ internal static class SessionExtensions
         {
             if (ex.Result.StatusCode == StatusCodes.BadEncodingLimitsExceeded)
             {
-                var chunkSize = (uint)((nodesToRead.Count / 3) >> 1) + 1u;
+                var chunkSize = ((nodesToRead.Count / 3) >> 1) + 1;
                 return nodeIds.Chunk(chunkSize).SelectMany(x => session.ReadNodeInfo(x));
             }
